pyproject.toml
[project]
name = "google-docs-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "fastapi==0.121.2",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.1",
    "google-auth-oauthlib==1.2.3",
    "googleapis-common-protos==1.72.0",
    "h11==0.16.0",
    "httplib2==0.31.0",
    "idna==3.11",
    "oauthlib==3.3.1",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pyparsing==3.2.5",
    "python-multipart==0.0.20",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "starlette==0.49.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
]
